Rename calculator core
