+++ mj-calc/src/modules/Calculator.tsx
